+++ frost-test/tests/server.rs
@@ -1,0 +1,21 @@
+use relay_server::{Call, Method, Response, Server};
+
+#[test]
+fn server_test() {
+    let mut server = Server::default();
+    {
+        let request = Method::POST.request(
+            "/".to_string(),
+            Default::default(),
+            "Hello!".as_bytes().to_vec(),
+        );
+        let response = server.call(request).unwrap();
+        let expected = Response::new(
+            200,
+            "OK".to_string(),
+            Default::default(),
+            Default::default(),
+        );
+        assert_eq!(response, expected);
+    }
+}
